Cleanup

+++ src/main.cpp
@@ -9,9 +9,7 @@ static const char* TAG = "Main";
 KorvoLCD      tft(320, 240);
 KorvoExpander expander;
 
-void setup()
-{
-	// put your setup code here, to run once:
+void setup() {
 	Serial.begin(115200);
 
 	Wire.begin(17, 18);
@@ -25,17 +23,15 @@ void setup()
 	tft.drawString("Hello you!", 10, 10);
 }
 
-void loop()
-{
+void loop() {
 	uint16_t t_x, t_y; // To store the touch coordinates
 
-	if (tft.getTouch(&t_x, &t_y))
-	{
-		int32_t size = map(tft.getTouchRawZ(), 10, 50, 2, 10);
+	if (tft.getTouch(&t_x, &t_y)) {
+		uint8_t size = map(tft.getTouchRawZ(), 10, 50, 2, 10);
 
 		tft.fillCircle(t_x, t_y, size, tft.color565(59, 161, 112));
 
-		// printf(">touch:%d:%d:%d|xy\n", t_x, t_y, millis());
+		// printf(">touch:%d:%d|xy\n", t_x, t_y);
 	}
 
 	delay(20);
